fix: resolve remaining MSVC compilation warnings in tests and source
- Fixed C4101 'unreferenced local variable' in `XLComments::setRichText` by catching `XLException const&` anonymously.
- Resolved C4244 `uint32_t` to `uint16_t` implicit conversion warnings in `XLTables::createTable` and `XLStreamReader::nextRow` by applying explicit integer truncations (`static_cast<uint16_t>`).
- Suppressed size conversion warnings across the PivotTable and Style facade modules when running under strict MSVC warning levels.

+++ OpenXLSX/sources/XLTables.cpp
@@ -107,8 +107,8 @@ namespace OpenXLSX
                                "\"/>"
                                "<tableColumns count=\"" +
                                std::to_string(colCount) + "\">";
-        for (uint32_t i = 0; i < colCount; ++i) {
-            std::string h = m_worksheet->cell(start.row(), start.column() + i).value().getString();
+        for (uint16_t i = 0; i < colCount; ++i) {
+            std::string h = m_worksheet->cell(start.row(), static_cast<uint16_t>(start.column() + i)).value().getString();
             if (h.empty()) h = "Column" + std::to_string(i + 1);
             tableXml += "<tableColumn id=\"" + std::to_string(i + 1) + "\" name=\"" + h + "\"/>";
         }
